+++ src/commandIds.ts
@@ -1,26 +1,31 @@
 const EXTENSION_PREFIX = 'sorbetto.';
 
 /**
+ * Run `bundle install`.
+ */
+export const BUNDLE_INSTALL_ID = `_${EXTENSION_PREFIX}bundleInstall`;
+
+/**
  * Copy Symbol to Clipboard.
  */
-export const COPY_SYMBOL_COMMAND_ID = `${EXTENSION_PREFIX}copySymbolToClipboard`;
+export const COPY_SYMBOL_ID = `${EXTENSION_PREFIX}copySymbolToClipboard`;
+
+/**
+ * Setup workspace.
+ */
+export const SETUP_WORKSPACE_ID = `${EXTENSION_PREFIX}setup.workspace`;
 
 /**
  * Show Sorbet Output panel.
  */
-export const SHOW_OUTPUT_COMMAND_ID = `${EXTENSION_PREFIX}showOutput`;
+export const SHOW_OUTPUT_ID = `_${EXTENSION_PREFIX}showOutput`;
 
 /**
  * Save package files. This command name is required by Sorbet itself.
  */
-export const SORBET_SAVE_PACKAGE_FILES_COMMAND_ID = 'sorbet.savePackageFiles';
+export const SORBET_SAVE_PACKAGE_FILES_ID = 'sorbet.savePackageFiles';
 
 /**
  * Restart Sorbet.
  */
-export const SORBET_RESTART_COMMAND_ID = `${EXTENSION_PREFIX}restart`;
-
-/**
- * Verify workspace.
- */
-export const VERIFY_WORKSPACE_COMMAND_ID = `${EXTENSION_PREFIX}verify.workspace`;
+export const SORBET_RESTART_ID = `${EXTENSION_PREFIX}restart`;